API Reference

Public Types

AntennaFieldRepresentations.AbsorbedSphericalExpansion — Type

AbsorbedSphericalExpansion{C<:Complex} <: AbstractSphericalExpansion

Store spherical vector wave expansion coefficients for an expansion of absorbed type

Fields

coefficients::Array{C,1} : array storing the expansion coefficients

AntennaFieldRepresentations.AbstractDipole — Type

Abstract type container type for HertzDipole and FitzgeraldDipole

AntennaFieldRepresentations.AbstractSphericalExpansion — Type

SphericalExpansions can be Radiating, Incident, or Absorbed

AntennaFieldRepresentations.AbstractSurfaceCurrentDensity — Type

Abstract type container type for Surface current densities

AntennaFieldRepresentations.FarfieldPattern — Type

Radiated far-field pattern sampled in angular domain. Sampling points lie on equiangular grid in ϕ and Gauß-Legendre grid in θ. Number and location of sampling points is defined by L. length(θ)=L+1 length(ϕ)=2L+2.

Fields:

k₀::Float64 : Wavenumber L::Int64 : Order of the representation Eθ::Array{<:Complex,2} : matrix storing the complex values for the θ-components Eϕ::Array{<:Complex,2} : matrix storing the complex values for the ϕ-components

AntennaFieldRepresentations.FitzgeraldDipole — Type

FitzgeraldDipole{T<:Number} <: AbstractDipole

Store information for a magnetic dipole

Fields

pos::SArray{Tuple{3},T,1,3} : position (can be real or complex)

dir::SArray{Tuple{3},T,1,3} : direction (must be complex to enable circular polarization)

mag::ComplexF64 : magnitude (for scalar excitation)

AntennaFieldRepresentations.HertzDipole — Type

HertzDipole{T<:Number} <: AbstractDipole

Store information for an electric dipole

Fields

pos::SArray{Tuple{3},T,1,3} : position (can be real or complex)

dir::SArray{Tuple{3},T,1,3} : direction (must be complex to enable circular polarization)

mag::ComplexF64 : magnitude (for scalar excitation)

AntennaFieldRepresentations.IncidentSphericalExpansion — Type

IncidentSphericalExpansion{C<:Complex} <: AbstractSphericalExpansion

Store spherical vector wave expansion coefficients for an expansion of incident type

Fields

coefficients::Array{C,1} : array storing the expansion coefficients

AntennaFieldRepresentations.MLFMMInteractionMatrix — Type

MLFMMInteractionMatrix{C} <: LinearMaps.LinearMap{C}

Represent interactions between array of dipole sources and array of dipole probes as abstract matrix

Fields

sourcestruct::AntennaFieldRepresentations.MLFMMSource

receivestruct::AntennaFieldRepresentations.MLFMMReceive

AntennaFieldRepresentations.PlaneWave — Type

Object for a single plane wave

Fields:

kvec::Array{<:Number,1} : k-vector determining the propagation direction. If complex valued: Evanescent components pol::Array{<:Number,1} : polarizatiion vector. If complex valued: elliptic polarization mag::Complex : Complex valued excitation amplitude

AntennaFieldRepresentations.PlaneWaveSpectrum — Type

Incident plane-wave spectrum sampled in angular domain. Sampling points lie on equiangular grid in ϕ and Gauß-Legendre grid in θ. Number and location of sampling points is defined by L. length(θ)=L+1 length(ϕ)=2L+2.

Fields:

k₀::Float64 : Wavenumber L::Int64 : Order of the representation Eθ::Array{Complex{Float64},2} : matrix storing the complex values for the θ-components Eϕ::Array{Complex{Float64},2} : matrix storing the complex values for the ϕ-components

AntennaFieldRepresentations.RadiatingSphericalExpansion — Type

RadiatingSphericalExpansion{C<:Complex} <: AbstractSphericalExpansion

Store spherical vector wave expansion coefficients for an expansion of radiating type

Fields

coefficients::Array{C,1} : array storing the expansion coefficients

Public Functions

AntennaFieldRepresentations.Wacker — Method

Wacker(S12, αin, [θvec, weightvec])

Return (receiving) spherical mode coefficients from measured S12 data with regular grid points on sphere. S12 is expected to have dimensions (L+1) × (2L+1) × 2 if no θvec or Jθ, JΦ, Jχ is given

AntennaFieldRepresentations.Wackerforward — Method

Wackerforward(α_inc,  β_aut, Jθ, Jϕ[, Jχ])

Return regularly sampled S12 measurement data on sphere.

AntennaFieldRepresentations.Wackerforward_ad — Method

Wackerforward_ad(α_inc, S21, Jθ, Jϕ, Jχ)

Return adjoint operator to Wackerforward(α_inc, β_aut, Jθ, Jϕ, Jχ).

AntennaFieldRepresentations.cdist — Method

cdist(R::Array{<:Number,1}) -> Number

Return √(R⋅R) without complex conjugation of R in the dot product. Corresponds to the norm operator for real valued vectors

AntennaFieldRepresentations.convertrepresentation — Function

convertrepresentation(newType<:Type{FieldRepresentation}, fieldobject<: FieldRepresentation, args)

Convert field representation into other type with different datastructure.

Field evaluations of the respective old and new field representations should be approximately equal in a common region of convergence.

AntennaFieldRepresentations.converttype — Function

converttype(newType<:Type{FieldRepresentation}, fieldobject<: FieldRepresentation)

Convert field representation into other type with same datastructure. Field avaluations of the respective old and new field representations will differ.

AntennaFieldRepresentations.efield — Method

efield(fieldobject<:FieldRepresentation, R, k₀::Number) -> [Ex;Ey;Ez]
efield(Vector{<:FieldRepresentation}, R, k₀::Number) -> [Ex;Ey;Ez]

Return the E-field vector (in cartesian coordinates) of the field representation at location R.

AntennaFieldRepresentations.ehfield — Method

ehfield(fieldobject<:FieldRepresentation, R, k₀::Number) -> [Ex;Ey;Ez], [Hx;Hy;Hz]
ehfield(Vector{<:FieldRepresentation}, R, k₀::Number) -> [Ex;Ey;Ez], [Hx;Hy;Hz]

Return the E-field and H-field vector (in cartesian coordinates) of the field representation at location R.

AntennaFieldRepresentations.elementtype — Function

elementtype(fieldobject<:FieldRepresentation)

Return base type of elements representing the fields in fieldobject. Default is ComplexF64

AntennaFieldRepresentations.hfield — Method

hfield(fieldobject<:FieldRepresentation, R, k₀::Number) -> [Hx;Hy;Hz]
hfield(Vector{<:FieldRepresentation}, R, k₀::Number) -> [Hx;Hy;Hz]

Return the H-field vector (in cartesian coordinates) of the field representation at location R.

AntennaFieldRepresentations.j_to_sℓm — Method

j_to_sℓm(j)

Convert single index j to multi-index s ℓ m

AntennaFieldRepresentations.reciprocaltype — Method

reciprocaltype(type::Type{<:AbstractElectromagneticFieldRepresentation}) -> Type{<:AbstractElectromagneticFieldRepresentation}

Return type of the field representation which is reciprocal to the input type.

AntennaFieldRepresentations.resample — Method

resample(pattern::PlaneWaveRepresentation, L::Integer, [orderθ::Integer=12, orderϕ::Integer=12]) -> newpattern::PlaneWaveRepresentation

Return pattern sampled according to order L using Lagrange interpolation of orders orderθ, orderϕ

AntennaFieldRepresentations.revertdirection — Method

revertdirection(pattern::PlaneWaveRepresentation) -> pattern_revert::PlaneWaveRepresentation

Return PlaneWaveRepresentation with reverted propagation directions. F(k) -> F(-k)

AntennaFieldRepresentations.rot_mat_zyz — Method

rot_mat_zyz(χ::Number, θ::Number, ϕ::Number)-> R::Matrix{eltype(χ, θ, ϕ)}

Return rotation matrix for cartesian vector components for a rotation around Euler angles χ, θ, ϕ.

AntennaFieldRepresentations.rotate — Method

rotate(fieldobject<:FieldRepresentation, χ::Number, θ::Number, ϕ::Number) -> rotatedfieldobject<:Fieldrepresentation
rotate(fieldobject<:FieldRepresentation; [χ=0.0],  [θ=0.0], [ϕ=0.0]) -> rotatedfieldobject<:Fieldrepresentation

Rotate a field representation around the Euler angles χ, θ, ϕ.

AntennaFieldRepresentations.samplingrule — Method

samplingrule(L::Integer) -> ( w::Array{Float64,1}, θ::Array{Float64,1}, ϕ::Array{Float64,1} )

Return sampling points for θ, ϕ, and θ-integration weights w for a far field or plane wave spectrum of order L.

AntennaFieldRepresentations.shiftrepresentation — Function

shiftrepresentation(fieldobject<:FieldRepresentation, R, k₀::Number) -> newobject<:Fieldrepresentation

Translate field representation into field representation of same type in translated coordinate system with new origin at R.

AntennaFieldRepresentations.sqrtdot — Method

sqrtdot(u::Array{<:Number,1}, v::Array{<:Number,1}) -> Number

Return √(u⋅v) without complex conjugation of u

AntennaFieldRepresentations.sℓm_to_j — Method

sℓm_to_j(s,ℓ,m)

Convert multi-index s ℓ m to single index j

AntennaFieldRepresentations.translate — Function

translate(fieldobject<:FieldRepresentation, R, k₀::Number) -> newobject<:Fieldrepresentation

Translate radiating field representation into incident field representation in translated coordinate system with new origin at R.

AntennaFieldRepresentations.transmission — Method

transmission(sender<:FieldRepresentation, receiver<:FieldRepresentation, k₀::Number) -> val::Complex

Return interaction between two field representations ( or source representations)

AntennaFieldRepresentations.udot — Method

dotu(u::Array{<:Number,1}, v::Array{<:Number,1}) -> Number

Return u⋅v without complex conjugation of u

Private Functions and Types

AntennaFieldRepresentations.EquiangularFarFieldPattern — Type

Far-field pattern sampled in angular domain. Sampling points lie on equiangular grid in ϕ and θ.

Fields:

k₀::Float64 Jθ::Integer : defines sampling density in θ: Δθ= 2π/Jθ Jϕ::Integer : defines sampling density in ϕ: Δϕ= 2π/Jϕ Eθ::Array{Complex{Float64},2} : matrix storing the complex values for the corresponding co polar plane wave Eϕ::Array{Complex{Float64},2} : matrix storing the complex values for the corresponding cross polar plane wave

AntennaFieldRepresentations.UnorthodoxSphericalExpansion — Type

UnorthodoxSphericalExpansion{C<:Complex} <: AbstractSphericalExpansion

Store spherical vector wave expansion coefficients for an expansion of absorbed type

Fields

coefficients::Array{C,1} : array storing the expansion coefficients

AntennaFieldRepresentations.F_sℓm_cartesian — Method

F_sℓm_cartesian(s,ℓ,m,c,R,k0) -> [Fx;Fy;Fz]

Return normalized vector spherical wave function at R in cartesian coordinates

AntennaFieldRepresentations.F_sℓm_cartesian_array — Method

F_sℓm_cartesian_array(s,ℓ,m,c,R,k0) -> [Fx;Fy;Fz]

Return normalized vector spherical wave function at R in cartesian coordinates

AntennaFieldRepresentations.F_sℓm_spherical — Method

F_sℓm_spherical(s,ℓ, m,r,ϑ,φ) -> [Fr;Fϑ;Fφ]

Return normalized vector spherical wave function at r,ϑ,φ in spherical coordinates

AntennaFieldRepresentations.F_sℓm_spherical_array — Method

F_sℓm_spherical_array(Jmaxxc,r,ϑ,φ, k0) -> Fr[1:Jmaxx], Fϑ[1:Jmaxx], Fφ[1:Jmaxx]

Return all spherical vector wave functions upt to Jmaxx at r, ϑ, φ in spherical coordinates

AntennaFieldRepresentations.F_sℓm_spherical_rzero — Method

F_sℓm_rzero(s,ℓ, m ,c,ϑ,φ, k0) -> Fr,Fϑ,Fφ

Catches special cases for r==0 Hansen p. 3245f.

AntennaFieldRepresentations.F_sℓm_thetapi — Method

F_sℓm_thetapi(s,ℓ,m,c,r,φ,k0) -> Fr,Fϑ,Fφ

Catches special cases for ϑ==pi Hansen p. 3245f.

AntennaFieldRepresentations.F_sℓm_thetazero — Method

F_sℓm_thetazero(s,ℓ,m,c,r,φ,k0)- > Fr,Fϑ,Fφ

Catches special cases for ϑ==0 Hansen p. 3245f.

AntennaFieldRepresentations.Greensdivdiv — Method

Greensdivdiv(R::SArray{Tuple{3},C,1,3}, dipoledir, k₀::Float64) -> SArray{Tuple{3},C,1,3}
Greensdivdiv(R::Array{<:Number,3}, dipoledir, k₀::Float64) -> Array{<:Number,3}

Return dyadic operator (I+1/k² ∇∇) g₀(r,r')

AntennaFieldRepresentations.Greensrot — Method

Greensrot(R::SArray{Tuple{3},C,1,3}, dipoledir, k₀::Float64) -> SArray{Tuple{3},C,1,3} Greensrot(R::Array{<:Number,3}, dipoledir, k₀::Float64) -> Array{<:Number,3}

Return dyadic operator ∇g₀(r,r') × I

AntennaFieldRepresentations.K_sℓm — Method

K_sℓm(s,ℓ,m,ϑ,φ) -> Kϑ,Kφ

Return farfield vector spherical wave function in spherical coordinates

AntennaFieldRepresentations.K_sℓm_array — Method

K_sℓm_array(Jmaxx,ϑ,φ) -> Kϑ[1:Jmaxx], Kφ[1:Jmaxx]

Return farfield for all vector spherical wave functions up to Jmaxx in spherical coordinates

AntennaFieldRepresentations.K_sℓm_thetapi — Method

K_sℓm_thetapi(s,ℓ,m,c,φ) -> Kϑ, Kφ

Catches special cases for ϑ==π Hansen p. 329f.

AntennaFieldRepresentations.K_sℓm_thetazero — Method

K_sℓm_thetazero(s,ℓ,m,c,φ) -> Kϑ, Kφ

Catches special cases for ϑ==0 Hansen p. 329f.

AntennaFieldRepresentations.Plm_deriv_array — Method

Plm_deriv_array(Lmax,m,x)

Return unnormalized Associated Legendre polynomials and their derivaives up to Lmax

AntennaFieldRepresentations._adjoint_aggregate! — Method

_adjoint_aggregate!(A::MLFMMSource, aggregationlist::Vector{Vector{Int}})

Perform adjoint operation (i.e., complex conjugate of transposed operation) of _adjoint_aggregate!

AntennaFieldRepresentations._adjoint_aggregate_leafnodes! — Method

_adjoint_aggregate_leafnodes!(A::MLFMMSource)

Perform the adjoint operation (i.e., conjugate of transposed operation) to _aggregate_leafnodes!

AntennaFieldRepresentations._adjoint_disaggregate! — Method

_adjoint_disaggregate!(A::MLFMMReceive, [y::AbstractVector])

Perform adjoint operation (i.e., complex conjugate of transposed operation) of _disaggregate!#

If no vector y is given, the content in A.bvector is used for adjoint disaggregation. Otherwise, A.bvector is overwritten by y before adjoint disaggregation.

AntennaFieldRepresentations._adjoint_disaggregate_leafnodes! — Method

_adjoint_disaggregate_leafnodes!(A::MLFMMReceive)

Perform ajoint operation (i.e., complex conjugate of transposed operation) of _disaggregate_leafnodes!

AntennaFieldRepresentations._adjoint_forward! — Method

_adjoint_forward!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive)

Store the result of the adjoint matrix vector product in sourcevector.xvector

If no vector y is given, receivestruct.bvector is used as tight vector. Otherwise, receivestruct.bvector is overwritten by the content of y before the operation is performed

AntennaFieldRepresentations._adjoint_transfer! — Method

_adjoint_transfer!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive)

Perform the adjoint operator (i.e., complex conjugate of transposed operator) of transfer!

AntennaFieldRepresentations._aggregate! — Method

_aggregate!(A::MLFMMSource, aggregationlist::Vector{Vector{Int}})

Aggregate A according to aggregationlist.

The aggregationlist specifies for each level which nodes shall be aggregated.

AntennaFieldRepresentations._aggregate_children! — Method

_aggregate_children!(A::MLFMMSource, parentnode)

Fill pattern storage of parentnode with aggregated pattern from all its children.

AntennaFieldRepresentations._aggregate_leafnodes! — Method

_aggregate_leafnodes!(A::MLFMMSource)

Fill storage for leaf node patterns due to excitation vector A.xvector.

AntennaFieldRepresentations._aggregate_to_minlevel! — Method

aggregatetominlevel!(A::MLFMMSource, [x::AbstractVector]; minaggregationlevel::Integer=0)

Aggregate A up to min_aggregationlevel.

AntennaFieldRepresentations._allocatenodepattern — Method

_allocatenodepattern(P::Type{S}, tree::MLFMMTrees.AbstractMLFMMTree, cutoffparameters::Vector{<:Integer}, [transferlist::Vector{Vector{Int}}, minlevelel::Int]) where {R<:Real, S<:Union{PlaneWaveSpectrum{Complex{R}}, FarfieldPattern{Complex{R}}}}

Allocate memory for intermediate representation of far field patterns for each node in tree.

If transferlist is provided, only node patterns which are relevant for translations are stored. If minlevel is provided, only node patterns up to minlevel are stored.

AntennaFieldRepresentations._collectPl! — Method

_collectPl!(Pℓstorage, Lmax,x)

Return Legendre polynomials up to Lmax with preallocated storage

AntennaFieldRepresentations._cutoffparameter — Method

_cutoffparameter(diameter, expectedaccuracy, k0)

Return cutoff parameter (i.e., maximum spherical mode index L) to represent sources inside sphere of given diameter with expected accuracy.

AntennaFieldRepresentations._disaggregate! — Method

_disaggregate!(receivestruct::MLFMMReceive)

Perform all disaggregations according to receivestruct.disaggregationlist. The receivestruct.disaggregationlist stores all nodes at each level which shall perform a disaggregation.

AntennaFieldRepresentations._disaggregate_children! — Method

_disaggregate_children!(receivestruct, parentnode)

Disaggregate pattern of parentnode to all its children and store the resulting patterns in receivestruct.nodespectra[child]

AntennaFieldRepresentations._disaggregate_leafnodes! — Method

_disaggregate_leafnodes!(receivestruct::MLFMMReceive)

Test leaf node patterns with all test functions of corresponding leafnode and store result in receivestruct.bvector

AntennaFieldRepresentations._extract_θ_row! — Method

_extract_θ_row!(storage, Ematr::Array{C,2}, iθrange, wθ::<:AbstractVector) where {C<:Complex}

Use interpolation weights wθ and correctly mapped range 'iθrange' to extract values for all plane waves in certain θ-cut from matrix of PlaneWaveRepresentation.

AntennaFieldRepresentations._forward! — Method

_forward!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive,[x])

Store the result of the matrix vector product in receivestruct.bvector

If no vector x is given, sourcestruct.xvector is used as excitation vector. Otherwise, sourcestruct.xvector is overwritten by the content of x before the operation is performed

AntennaFieldRepresentations._getpoints — Method

_getpoints(basisfunctions)

Return list of SVector{3} representing the reference (i.e., center) locations of the basisfunctions

AntennaFieldRepresentations._getsourcefunctions — Method

_getsourcefunctions(basisfunctions)

Return list of individual source functions

AntennaFieldRepresentations._initialize_aggregationlist — Method

_initialize_aggregationlist(TXtree::MLFMMTrees.AbstractMLFMMTree, transferlist)

Return a list of nodes per level which shall be aggregated to be able to perform all transfers provided in transferlist.

AntennaFieldRepresentations._initialize_disaggregationlist — Method

_initialize_disaggregationlist(RXtree::MLFMMTrees.AbstractMLFMMTree, transferlist)

Return a list of nodes per level which shall be disaggregated to be able to perform all disaggregations from nodes which receive transfers provided in transferlist.

AntennaFieldRepresentations._initialize_transferlist — Method

_initialize_transferlist(TXtree::MLFMMTrees.AbstractMLFMMTree, RXtree::MLFMMTrees.AbstractMLFMMTree; num_bufferboxes::Integer=1)

For each node in RXtree, return a list of all transmit nodes in TXtree which shall perform a transfer to the receive node.

AntennaFieldRepresentations._initialize_tree — Method

_initialize_tree(points::AbstractArray{SVector{3,R}}, minhalfsize::R) where{R<:Real}

Assemble 'MLFMMTree' based on point cloud and half length of leave boxes

AntennaFieldRepresentations._initializebasisfunctionfarfields — Method

_initializebasisfunctionfarfields(tree::MLFMMTrees.AbstractMLFMMTree, basisfunctions::S, cutoffparameters::Vector{I}, k0::R, verbose::Bool) where {R<:Real, S<:AbstractSurfaceCurrentDensity, I<:Int}

Return list of far fields for every basis function on leaf level.

AntennaFieldRepresentations._initializebasisfunctionweightingpatterns — Method

_initializebasisfunctionweightingpatterns(tree::MLFMMTrees.AbstractMLFMMTree, functionspace,
cutoffparameters::Vector{<:Integer}, k0::R, verbose::Bool) where{R<:Real}

Return list of far fields for every test function on leaf level stored in reverted direction feasible for testing incident plane wave spectra.

AntennaFieldRepresentations._initializelevelcutoffparameters — Method

_initializelevelcutoffparameters(tree::MLFMMTrees.AbstractMLFMMTree, expectedaccuracy, wavenumber)

Return list of cutoff parameters for each level in tree to represent the node farfields with expected accuracy.

AntennaFieldRepresentations._initializelevelinterpolators — Method

_initializelevelinterpolators(R::Type{<:Real}, tree::MLFMMTrees.AbstractMLFMMTree, orderθ::Integer, orderϕ::Integer,cutoffparameters::Vector{<:Integer}; minlevel::Int=0)

Return list of interpolators to aggregate from each level to the respective parent level.

AntennaFieldRepresentations._initializetransferplan — Method

_initializetransferplan(transferlist::Vector{Vector{Int}}, sourcestruct::MLFMMSource, RXtree::MLFMMTrees.AbstractMLFMMTree;  ) where{P<:Type{<:AbstractTransfer}}

For each planned transfer, initialize and store a transferobject which holds all rewuired information to calculate the corresponding far field transfer.

Keywordarguments:

transfertype::P=PlannedTransfer{Complex{R}} : specify type of the initialized tranfer object

AntennaFieldRepresentations._phaseshiftmatrix! — Method

_phaseshiftmatrix!(storagematrix, R, k₀)

In-place calculation of phaseshiftmatrix which needs to be element-wise multiplied to farfield to shift the farfield to new coordinate origin at R. Assumes that size(storagematrix) mathces the matrices in PlaneWaveRepresentation.

AntennaFieldRepresentations._transfer! — Method

_transfer!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive)

Perform all required transfers between the sourcestruct and the receivestruct.

AntennaFieldRepresentations._transpose_aggregate! — Method

_transpose_aggregate!(A::MLFMMSource, aggregationlist::Vector{Vector{Int}})

Perform transpose operation of _aggregate!

AntennaFieldRepresentations._transpose_aggregate_children! — Method

_transpose_aggregate_children!(A, parentnode)

Perform transposed operation to _aggregate_children!

AntennaFieldRepresentations._transpose_aggregate_leafnodes! — Method

_transpose_aggregate_leafnodes!(A::MLFMMSource)

Perform the transpose operation to _aggregate_leafnodes!

AntennaFieldRepresentations._transpose_disaggregate! — Method

_transpose_disaggregate!(A::MLFMMReceive, [y::AbstractVector])

Perform transposed operation of `_disaggregate!`

If no vector y is given, the content in A.bvector is used for transposed disaggregation. Otherwise, A.bvector is overwritten by y before transposed disaggregation.

AntennaFieldRepresentations._transpose_disaggregate_children! — Method

_transpose_disaggregate_children!(A, parentnode)

Perform transposed operation to disaggrgate_children!

AntennaFieldRepresentations._transpose_disaggregate_leafnodes! — Method

_transpose_disaggregate_leafnodes!(A::MLFMMReceive)

Perform transposed operation to disaggregate_leafnodes!

AntennaFieldRepresentations._transpose_forward! — Method

_transpose_forward!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive, [y])

Store the result of the transposed matrix vector product in sourcevector.xvector

If no vector y is given, receivestruct.bvector is used as tight vector. Otherwise, receivestruct.bvector is overwritten by the content of y before the operation is performed

AntennaFieldRepresentations._transpose_transfer! — Method

_transpose_transfer!(sourcestruct::MLFMMSource, receivestruct::MLFMMReceive)

Perform the transposed operator to transfer!

AntennaFieldRepresentations.collectPl — Method

collectPl(Lmax,x)

Return Legendre polynomials up to Lmax

AntennaFieldRepresentations.extract_single_planewave — Method

extract_single_planewave(Eθmatr::Array{<:Complex,2}, Eϕmatr::Array{<:Complex,2}, iθrange, wθ, iϕrange, wϕ ) -> Eθ, Eϕ

Use interpolation weights wθ, wϕ and correctly mapped ranges 'iθrange', iϕrange to extract Eθ- and Eϕ- value for single plane wave from PlaneWaveRepresentation.

AntennaFieldRepresentations.find_next_smaller_θind — Method

find_next_smaller_θind(θvec::Array{<:Real,1}, θnew<:Real) -> i::Integer

Return index of largest entry in ascendingly ordered θvec which is still smaller than θnew

AntennaFieldRepresentations.find_next_smaller_ϕind — Method

find_next_smaller_ϕind(Δϕ<:Real, ϕnew<:Real)-> i::Integer

Return index of largest entry in ascendingly ordered ϕvec which is still smaller than ϕnew

AntennaFieldRepresentations.g₀ — Method

g₀(R, k₀) -> Number

Scalar Greens function exp(-jk₀R)/(4πR)

AntennaFieldRepresentations.interpolate_single_planewave — Method

interpolate_single_planewave(θnew::Real, ϕnew::Real, pattern::PlaneWaveRepresentation, ::AbstractInterpolationStrategy)

Interpolate single plane wave into direction θnew, ϕnew from given PlaneWaveRepresentation using plynomial (Legendre) interpolation of order orderθ in θ and orderϕ in ϕ.

AntennaFieldRepresentations.lagrange_barycentric_weights — Method

lagrange_barycentric_weights(X::Array{<:Number,1}) -> w::Array{<:Number,1}

Return barycentric weights for calculating Lagrange interpolation weights

Extended help

The barycentric weights are needed to evaluate the "second (true) form of the barycentric formula" for the Lagrange interpolation. See J.-P. Berrut and L. N. Trefethen: "Barycentric Lagrange Interpolation", SIAM Review, Vol. 46, No. 3, pp. 501-517

AntennaFieldRepresentations.lagrange_interpolation_weights — Method

lagrange_interpolation_weights(X::Array{::Number,1},Y::Array{::Number,1},t::Number) -> w::Array{Flota64}

Return interpolation weights wᵢ for the calculation f(t)= ∑ᴺᵢ₌₁ wᵢ⋅ f(xᵢ)

Extended help

The interpolation weights are obtained using the "second (true) form of the barycentric formula". See J.-P. Berrut and L. N. Trefethen: "Barycentric Lagrange Interpolation", SIAM Review, Vol. 46, No. 3, pp. 501-517

AntennaFieldRepresentations.legendre_deps_array — Method

legendre_deps_array(m , Lmax, ϑ)

Return the theta dependencies occuring in the spherical wave expansion

AntennaFieldRepresentations.map_θrange! — Method

map_θrange!(iθrange, Nθ)

Map the index range iθrange to values between -Nθ+1 and Nθ.

AntennaFieldRepresentations.map_ϕrange! — Method

map_ϕrange!(iϕrange, Nθ)

Map the index range iϕrange to values between 1 and Nϕ.

AntennaFieldRepresentations.sphPlm_deriv_array — Method

sphPlm_deriv_array(Lmax,m,x)

Return normalized Associated Legendre polynomials and their derivaives up to Lmax sphPlm= 1/√(2π) √((2ℓ+1)(ℓ-m)!/(2(ℓ+m)!) Plm

From T. Limpanuparb, J. Milthorpe: "Associated Legendre Polynomials and Spherical Harmonics Computation for Chemistry Applications", arXiv 2014

AntennaFieldRepresentations.αinc_PW — Method

αinc_PW(L::Integer)

Return incident spherical mode coefficients up to mode order L of an x-polarized plane wave traveling in negative z-direction. Compare Hansen: "Spherical Near-Field Measurements" Appendix A1.6

AntennaFieldRepresentations.θsequence — Method

θsequence(θvec::Array{<:Real, 1}, iθrange::Array{<:Integer,1})

Truncate and/or periodically extend θvec to create θsequence matching the index range in iθrange.

AntennaFieldRepresentations.MLFMMTrees.getboundingbox — Method

getboundingbox(points::Vector{SVector{D, F}})

Returns halfsize and center of bounding box of points. The halfsize is the half of the length of the edge of the bounding box.